reasoning assignments all parts

+++ src/server/db/schema/reasoningAssignment.ts
@@ -1,4 +1,4 @@
-import { boolean, integer, pgTableCreator, timestamp, varchar, text } from "drizzle-orm/pg-core";
+import { boolean, integer, pgTableCreator, timestamp, varchar, text, decimal } from "drizzle-orm/pg-core";
 import { DATABASE_PREFIX as prefix } from "@/lib/constants";
 import { classrooms } from "./classroom";
 import { users } from "./user";
@@ -56,7 +56,7 @@ export const reasoningAssignmentAttempts = pgTable(
   {
     id: varchar("id", { length: 21 }).primaryKey(),
     assignmentId: varchar("assignment_id", { length: 21 }).references(() => reasoningAssignments.id),
-    score: integer("score"),
+    score: decimal("score"),
     submittedAt: timestamp("submitted_at", { mode: "date" }),
     userId: varchar("user_id", { length: 21 }).references(() => users.id),
     createdAt: timestamp("created_at").defaultNow().notNull(),
